+++ cmd/root.go
@@ -4,7 +4,7 @@ import (
 	"github.com/gin-gonic/gin"
 	"github.com/spf13/cobra"
 	"github.com/tass-io/scheduler/pkg/http"
-	"github.com/tass-io/scheduler/pkg/init"
+	initial "github.com/tass-io/scheduler/pkg/initial"
 	_ "github.com/tass-io/scheduler/pkg/tools/k8sutils"
 	_ "github.com/tass-io/scheduler/pkg/tools/log"
 	_ "github.com/tass-io/scheduler/pkg/workflow"
@@ -14,7 +14,7 @@ var rootCmd = &cobra.Command{
 	Use:   "scheduler",
 	Short: "scheduler",
 	Long:  "scheduler",
-	Run:   func(cmd *cobra.Command, args []string) {
+	Run: func(cmd *cobra.Command, args []string) {
 		r := gin.Default()
 		http.RegisterRoute(r)
 		r.Run()
@@ -27,5 +27,5 @@ func Execute() error {
 
 func init() {
 	rootCmd.AddCommand(versionCmd)
-	rootCmd.AddCommand(init.InitCmd)
+	rootCmd.AddCommand(initial.InitCmd)
 }
